change search string when looking for last clerk/cu
Before this change, the bot would look for the last clerk/CU to
edit the SPI page until it finds an edit summary that has the
substring "archiv", "moving", or "moved". This would sometimes
match against edit summaries that don't have to do with SPI
case moving.

For example, this edit summary:
https://en.wikipedia.org/w/index.php?title=Wikipedia:Sockpuppet_investigations/SimoooIX&diff=prev&oldid=1191010313

Produced this result:
https://en.wikipedia.org/w/index.php?title=Wikipedia:Sockpuppet_investigations/SPI/Cases&oldid=1192337359

(Notice there is no clerk/CU listed, even though Callanecc should
have been listed there.)

This change changes it so that it matches against the substring
"moving case", which is used by the spihelper script when moving
case sections out, e.g.
https://en.wikipedia.org/w/index.php?title=Wikipedia:Sockpuppet_investigations/NagalimNE&diff=prev&oldid=1192337834

This is there because moving a case out of an SPI and into another
SPI is effectively like archiving. Whatever new SPI case is started
after that action will be fresh.