Add support for IP range case names and make more robust against edge cases
This addresses two cases that produced unexpected behavior for the bot:

(1) IP range case names: the script does `.split('/')[1]` on the full page name
to get the name of the case, but if the name is an IP range, it will have two
forward slash characters (e.g. 'Wikipedia:Sockpuppet investigations/0.0.0.0/0').
In this case, the bot previously only grabbed the text between the two slashes,
resulting in a link to "Wikipedia:Sockpuppet investigations/0.0.0.0" when it
should be "Wikipedia:Sockpuppet investigations/0.0.0.0/0". Modified the logic
so that now it will grab the entire text after the first slash character.

(2) If there are no forward slash characters at all in the case page, as happened
for https://en.wikipedia.org/w/index.php?title=Wikipedia:Sockpuppet_investigations/Nv7801&action=history
then the bot would completely give up due to an uncaught IndexError. This will
no longer happen now: we explicitly check if the case name starts with
"Wikipedia:Sockpuppet investigations/", and if that is not true, then we will
put the entire page name as the link name.