Differentiate "cumoreinfo" and "moreinfo"
Currently, the SPI table bot script does not differentiate between
"cumoreinfo" and "moreinfo" in SPI case statuses. "cumoreinfo" is
specifically when more information is requested to determine whether
to use CheckUser, while "moreinfo" is when more information is
requested in a non-CheckUser case.

The script determines case status by looking at the categories for a
SPI case. However, "cumoreinfo" is not tied with a specific category.
Instead, it can only be deduced by looking for the presence of the
'SPI cases requesting more information' and the 'SPI cases for
pre-CheckUser review' together.