Change so that cases with the 'new' status are always displayed
Before, if a case was in the green 'checked' status, that would cause
any new case with the 'new' status not to show up in the table. This
adjusts the behavior to always display cases with the 'new' status
regardless of what other cases might be in the table.

This is desirable because cases with the 'new' status might be the
easy ones that can be knocked out quickly (and if they aren't, then
they should be moved to another status accordingly).

Example case where this would be desirable:
https://en.wikipedia.org/w/index.php?oldid=1324061675